pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-socket_tuner"
version = "0.1.0"
description = "Socket Tuner environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.80.0",
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "grpcio-tools>=1.80.0",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
    "protobuf>=6.33.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m socket_tuner.server.app
server = "socket_tuner.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["socket_tuner", "socket_tuner.server"]
package-dir = { "socket_tuner" = ".", "socket_tuner.server" = "server" }
